Add logic to ensure that short DOIs are resolved correctly
An alternative would be to encode short DOIs as:
`"http://shortdoi.org/" + encodeURIComponent(doi)`

But this approach seems less prone to breakage on the part of doi.org

I'm not sure what to do with the `_mapTag` function in `openurl.js` when it is passed a DOI item. This is mostly a concern for the CrossRef locator engine, which currently breaks when it passes a short DOI to the doi.org resolver. Shoul I add a conditional to `_mapTag` to look for DOIs in `data` and then treat long and short DOIs appropriately? Will being passed unescaped `/` break things if they are present in an OpenURL object (so that I should extract just the part after "10/" in the short DOI)?