Add successCodes argument to translator doGet/doPost
In 2652fac24 we started failing the translator if doGet/doPost returned
a >=400 status code, but at least one translator (Primo) relied on doGet
continuing after a 404. This allows translators to specify the status
codes that are allowed, similar to the same argument in
Zotero.HTTP.request().

(We'll clean up the signature at some point or just get rid of these
functions, but this is a quick fix for Primo.)