Restart after migrating legacy locale prefs
DTDs are loaded based on the initial state of `intl.locale.requested`,
so if `intl.locale.matchOS` was false, we need to restart after
migrating that to keep part of the UI from being in the wrong locale.
(Before zotero/zotero-standalone-build@ced8a0fa63, it would've been
English. After it would probably be the OS locale.)

https://forums.zotero.org/discussion/79673/zotero-5-0-77-beta-has-no-localization-anymore