Remove "Toggle Tag Selector" command from Shortcuts preference pane
Cmd/Ctrl-Shift-T is now used for reopening a closed tab

This is now a View menu option, so macOS users can assign an
app-specific shortcut key for it. For other platforms that can't do
that, Zutilo could add it, and it will be an option if we add general
key-binding configurability in a future version.

Closes #2281