Prevent duplicate collections during folder import
Improves the importer so that if collection with a matching name already
exists at given level (from previous import or otherwise), it will be
used instead of creating a duplicate with the same name.

Also code has been refactored to avoid looking up collections by path as
it is now redundant.