Copy related items between libraries
If all related item relations are within the drag set, relations are
copied without prompting the user. But if items in the drag set are
related to items outside the drag set, the user is asked whether to
additionally copy the related items and preserve the relations in the
target library.

The wording of the dialog text could use some work.