Item.setType: Extract fields, creators, archive IDs
Plus two fixes:

1. Fix Extra key/value regex

I'm fairly sure the '-' wasn't meant to be interpreted as a range -
space through underscore (0020 through 005F) covers punctuation and
uppercase, but not lowercase, letters, and it doesn't seem like a
meaningful range here. Since it contains the colon character (003A), it
caused the regex to eat up everything up to the *last* colon in a line
containing multiple colons, so
  arXiv ID: arXiv:12345
would be split into ['arXiv ID: arXiv', '12345'] instead of the correct
['arXiv ID', 'arXiv:12345'].

2. Fix extractExtraFields creator type comparison

getCreators() returns internal creator representations, so `creatorType`
isn't a valid field and the method wasn't behaving as the
comment ("Ignore if there are any creators of this type on the item
already") said it would.